src/br/com/joaomiguelg/domain/Produto.java
package br.com.joaomiguelg.domain;

public class Produto {

    private Long id;
    private String nome;
    private String preco;

    public String getPreco() {
        return preco;
    }

    public void setPreco(String preco) {
        this.preco = preco;
    }


    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }
}
